attempt to fix x86 vector-reduction splitting
