+++ src/CodeGen_LLVM.cpp
@@ -4009,7 +4009,7 @@ void CodeGen_LLVM::visit(const VectorReduce *op) {
     codegen_vector_reduce(op, Expr());
 }
 
-void CodeGen_LLVM::codegen_vector_reduce(const VectorReduce *op, const Expr &init) {
+Expr CodeGen_LLVM::split_vector_reduce(const VectorReduce *op, const Expr &init) const {
     Expr val = op->value;
     const int output_lanes = op->type.lanes();
     const int native_lanes = native_vector_bits() / op->type.bits();
@@ -4049,8 +4049,7 @@ void CodeGen_LLVM::codegen_vector_reduce(const VectorReduce *op, const Expr &ini
             equiv = max(equiv, init);
         }
         equiv = cast(op->type, equiv);
-        equiv.accept(this);
-        return;
+        return equiv;
     }
 
     if (op->type.is_bool() && op->op == VectorReduce::And) {
@@ -4061,8 +4060,7 @@ void CodeGen_LLVM::codegen_vector_reduce(const VectorReduce *op, const Expr &ini
         if (init.defined()) {
             equiv = min(equiv, init);
         }
-        equiv.accept(this);
-        return;
+        return equiv;
     }
 
     if (elt == Float(16) && upgrade_type_for_arithmetic(elt) != elt) {
@@ -4072,8 +4070,7 @@ void CodeGen_LLVM::codegen_vector_reduce(const VectorReduce *op, const Expr &ini
             equiv = binop(equiv, init);
         }
         equiv = cast(op->type, equiv);
-        equiv.accept(this);
-        return;
+        return equiv;
     }
 
     if (output_lanes == 1) {
@@ -4172,8 +4169,7 @@ void CodeGen_LLVM::codegen_vector_reduce(const VectorReduce *op, const Expr &ini
             if (initial_value.defined()) {
                 equiv = binop(initial_value, equiv);
             }
-            equiv.accept(this);
-            return;
+            return equiv;
         }
     }
 
@@ -4196,8 +4192,7 @@ void CodeGen_LLVM::codegen_vector_reduce(const VectorReduce *op, const Expr &ini
             equiv = binop(equiv, init);
         }
         equiv = common_subexpression_elimination(equiv);
-        equiv.accept(this);
-        return;
+        return equiv;
     }
 
     if (factor > 2 && ((factor & 1) == 0)) {
@@ -4229,8 +4224,7 @@ void CodeGen_LLVM::codegen_vector_reduce(const VectorReduce *op, const Expr &ini
             equiv = binop(equiv, init);
         }
         equiv = common_subexpression_elimination(equiv);
-        codegen(equiv);
-        return;
+        return equiv;
     }
 
     // Extract each slice and combine
@@ -4244,8 +4238,14 @@ void CodeGen_LLVM::codegen_vector_reduce(const VectorReduce *op, const Expr &ini
         }
     }
     equiv = common_subexpression_elimination(equiv);
-    codegen(equiv);
-}  // namespace Internal
+    return equiv;
+}
+
+void CodeGen_LLVM::codegen_vector_reduce(const VectorReduce *op, const Expr &init) {
+    Expr equiv = split_vector_reduce(op, init);
+    equiv.accept(this);
+    return;
+}
 
 void CodeGen_LLVM::visit(const Atomic *op) {
     if (!op->mutex_name.empty()) {
