+++ src/X86Optimize.h
@@ -5,6 +5,7 @@
  * Tools for optimizing IR for x86.
  */
 
+#include "CodeGen_LLVM.h"
 #include "Expr.h"
 #include "Target.h"
 
@@ -12,9 +13,7 @@ namespace Halide {
 namespace Internal {
 
 /** Perform vector instruction selection, inserting VectorIntrinsic nodes. */
-Stmt optimize_x86_instructions(Stmt s, const Target &t);
-
-Target complete_x86_target(Target t);
+Stmt optimize_x86_instructions(Stmt stmt, const Target &target, const CodeGen_LLVM *codegen);
 
 }  // namespace Internal
 }  // namespace Halide
